Separated Output from Script
Created an Output Script file and a Script file.

Output file contains a single statement that, when run, will source the Script file.

Script file contains all the code that will be run, including the output statements.

Also, I changed a few things in the script and added to my report.